Refactor PDF metadata extraction
