+++ API/Services/BookService.cs
@@ -13,6 +13,7 @@ using API.Entities;
 using API.Entities.Enums;
 using API.Extensions;
 using API.Services.Tasks.Scanner.Parser;
+using API.Helpers;
 using Docnet.Core;
 using Docnet.Core.Converters;
 using Docnet.Core.Models;
@@ -70,6 +71,8 @@ public class BookService : IBookService
     private static readonly RecyclableMemoryStreamManager StreamManager = new ();
     private const string CssScopeClass = ".book-content";
     private const string BookApiUrl = "book-resources?file=";
+    private readonly PdfMetadataExtractor _pdfMetadataExtractor;
+
     public static readonly EpubReaderOptions BookReaderOptions = new()
     {
         PackageReaderOptions = new PackageReaderOptions
@@ -85,6 +88,7 @@ public class BookService : IBookService
         _directoryService = directoryService;
         _imageService = imageService;
         _mediaErrorService = mediaErrorService;
+        _pdfMetadataExtractor = new PdfMetadataExtractor(_logger, _mediaErrorService);
     }
 
     private static bool HasClickableHrefPart(HtmlNode anchor)
@@ -426,189 +430,6 @@ public class BookService : IBookService
         }
     }
 
-    private int FindInBuffer(byte[] buffer, int bufLen, byte[] match)
-    {
-        var maxPos = bufLen - match.Length;
-        for (var pos = 0; pos<=maxPos; ++pos)
-        {
-            var found = true;
-            for (var ch = 0; ch<match.Length; ++ch)
-            {
-                if (buffer[pos+ch] != match[ch])
-                {
-                    found = false;
-                    break;
-                }
-            }
-            if (found)
-            {
-                return pos;
-            }
-        }
-        return -1;
-    }
-
-    private string? GetTextFromXmlNode(XmlDocument doc, XmlNamespaceManager ns, string path)
-    {
-        return (doc.DocumentElement?.SelectSingleNode(path + "//rdf:li", ns)
-            ?? doc.DocumentElement?.SelectSingleNode(path, ns))?.InnerText;
-    }
-
-    private float? GetFloatFromXmlNode(XmlDocument doc, XmlNamespaceManager ns, string path)
-    {
-        var text = GetTextFromXmlNode(doc, ns, path);
-        if (string.IsNullOrEmpty(text)) return null;
-
-        return float.Parse(text);
-    }
-
-    private string? GetListFromXmlNode(XmlDocument doc, XmlNamespaceManager ns, string path)
-    {
-        var nodes = doc.DocumentElement?.SelectNodes(path+"//rdf:li", ns);
-        if (nodes == null) return null;
-        var list = new StringBuilder();
-        foreach (XmlNode n in nodes)
-        {
-            if (list.Length > 0)
-            {
-                list.Append(",");
-            }
-            list.Append(n.InnerText);
-        }
-        return list.Length > 0 ? list.ToString() : null;
-    }
-
-    private DateTime? GetDateTimeFromXmlNode(XmlDocument doc, XmlNamespaceManager ns, string path)
-    {
-        var text = GetTextFromXmlNode(doc, ns, path);
-        if (text == null) return null;
-
-        return DateTime.Parse(text);
-    }
-
-    private ComicInfo? GetPdfComicInfoFromMetadata(string metadata, string filePath)
-    {
-        var metaDoc = new XmlDocument();
-        metaDoc.LoadXml(metadata);
-
-        var ns = new XmlNamespaceManager(metaDoc.NameTable);
-        ns.AddNamespace("rdf", "http://www.w3.org/1999/02/22-rdf-syntax-ns#");
-        ns.AddNamespace("dc", "http://purl.org/dc/elements/1.1/");
-        ns.AddNamespace("calibreSI", "http://calibre-ebook.com/xmp-namespace-series-index");
-        ns.AddNamespace("calibre", "http://calibre-ebook.com/xmp-namespace");
-        ns.AddNamespace("pdfx", "http://ns.adobe.com/pdfx/1.3/");
-        ns.AddNamespace("prism", "http://prismstandard.org/namespaces/basic/2.0/");
-        ns.AddNamespace("xmp", "http://ns.adobe.com/xap/1.0/");
-
-        var info = new ComicInfo();
-        var publicationDate = GetDateTimeFromXmlNode(metaDoc, ns, "//dc:date")
-            ?? GetDateTimeFromXmlNode(metaDoc, ns, "//xmp:createdate");
-        if (publicationDate != null)
-        {
-            info.Year  = publicationDate.Value.Year;
-            info.Month = publicationDate.Value.Month;
-            info.Day   = publicationDate.Value.Day;
-        }
-
-        info.Summary   = GetTextFromXmlNode(metaDoc, ns, "//dc:description") ?? String.Empty;
-        info.Publisher = GetTextFromXmlNode(metaDoc, ns, "//dc:publisher") ?? String.Empty;
-        info.Writer    = GetListFromXmlNode(metaDoc, ns, "//dc:creator") ?? String.Empty;
-        info.Title     = GetTextFromXmlNode(metaDoc, ns, "//dc:title") ?? String.Empty;
-        info.Genre     = GetListFromXmlNode(metaDoc, ns, "//dc:subject") ?? String.Empty;
-        info.LanguageISO = ValidateLanguage(GetTextFromXmlNode(metaDoc, ns, "//dc:language"));
-
-        info.Isbn = GetTextFromXmlNode(metaDoc, ns, "//pdfx:isbn") ?? GetTextFromXmlNode(metaDoc, ns, "//prism:isbn") ?? String.Empty;
-        if (!ArticleNumberHelper.IsValidIsbn10(info.Isbn) && !ArticleNumberHelper.IsValidIsbn13(info.Isbn))
-        {
-            _logger.LogDebug("[BookService] {File} has invalid ISBN number", filePath);
-            info.Isbn = String.Empty;
-        }
-
-        info.UserRating = GetFloatFromXmlNode(metaDoc, ns, "//calibre:rating") ?? 0.0f;
-        info.TitleSort  = GetTextFromXmlNode(metaDoc, ns, "//calibre:title_sort") ?? String.Empty;
-        info.Series     = GetTextFromXmlNode(metaDoc, ns, "//calibre:series/rdf:value") ?? String.Empty;
-        info.SeriesSort = info.Series;
-        info.Volume     = Convert.ToInt32(GetFloatFromXmlNode(metaDoc, ns, "//calibreSI:series_index") ?? 0.0f).ToString();
-
-        // If this is a single book and not a collection, set publication status to Completed
-        if (string.IsNullOrEmpty(info.Volume) && Parser.ParseVolume(filePath, LibraryType.Manga).Equals(Parser.LooseLeafVolume))
-        {
-            info.Count = 1;
-        }
-
-        var hasVolumeInSeries = !Parser.ParseVolume(info.Title, LibraryType.Manga)
-            .Equals(Parser.LooseLeafVolume);
-
-        if (string.IsNullOrEmpty(info.Volume) && hasVolumeInSeries && (!info.Series.Equals(info.Title) || string.IsNullOrEmpty(info.Series)))
-        {
-            // This is likely a light novel for which we can set series from parsed title
-            info.Series = Parser.ParseSeries(info.Title, LibraryType.Manga);
-            info.Volume = Parser.ParseVolume(info.Title, LibraryType.Manga);
-        }
-
-        ComicInfo.CleanComicInfo(info);
-
-        return info;
-    }
-
-    private ComicInfo? GetPdfComicInfo(string filePath)
-    {
-        try
-        {
-            const int chunkSize = 4096;
-            const int overlap = 16;
-            var stream = File.OpenRead(filePath);
-            var buffer = new byte[chunkSize + overlap];
-            var bytesAvailable = 0;
-            var hasMetaData = false;
-            var meta = new byte[0];
-            while (!hasMetaData)
-            {
-                var bytesRead = stream.Read(buffer, bytesAvailable, chunkSize);
-                if (bytesRead == 0) break;
-                bytesAvailable += bytesRead;
-                var found = FindInBuffer(buffer, bytesAvailable, Encoding.UTF8.GetBytes("<x:xmpmeta"));
-                if (found >= 0)
-                {
-                    meta = buffer[found..bytesAvailable];
-                    hasMetaData = true;
-                    break;
-                }
-                else
-                {
-                    var ovl = Math.Min(overlap, bytesAvailable);
-                    Buffer.BlockCopy(buffer, bytesAvailable - ovl, buffer, 0, ovl);
-                    bytesAvailable = ovl;
-                }
-            }
-            while ((bytesAvailable = stream.Read(buffer, 0, chunkSize)) > 0 || hasMetaData)
-            {
-                hasMetaData = false;
-                if (bytesAvailable > 0) {
-                    byte[] newMeta = new byte[meta.Length + bytesAvailable];
-                    Buffer.BlockCopy(meta, 0, newMeta, 0, meta.Length);
-                    Buffer.BlockCopy(buffer, 0, newMeta, meta.Length, bytesAvailable);
-                    meta = newMeta;
-                }
-                var found = FindInBuffer(meta, meta.Length, Encoding.UTF8.GetBytes("</x:xmpmeta>"));
-                if (found >= 0)
-                {
-                    var metadata = meta[0..(found + "</x:xmpmeta>".Length)];
-                    return GetPdfComicInfoFromMetadata(Encoding.UTF8.GetString(metadata), filePath);
-                }
-            }
-            return null; // Unterminated metadata
-        }
-        catch (Exception ex)
-        {
-            _logger.LogWarning(ex, "[GetComicInfo] There was an exception parsing PDF metadata");
-            _mediaErrorService.ReportMediaIssue(filePath, MediaErrorProducer.BookService,
-                "There was an exception parsing PDF metadata", ex);
-        }
-
-        return null;
-    }
-
     private ComicInfo? GetEpubComicInfo(string filePath)
     {
         try
@@ -771,7 +592,7 @@ public class BookService : IBookService
 
         if (Parser.IsPdf(filePath))
         {
-            return GetPdfComicInfo(filePath);
+            return _pdfMetadataExtractor.GetComicInfo(filePath);
         }
         else
         {
@@ -881,7 +702,7 @@ public class BookService : IBookService
         return (year, month, day);
     }
 
-    private static string ValidateLanguage(string? language)
+    public static string ValidateLanguage(string? language)
     {
         if (string.IsNullOrEmpty(language)) return string.Empty;
